+++ CHANGELOG.md
@@ -2,11 +2,13 @@ Version 1.1.1-SNAPSHOT:
 - add event-driven LDAP engine support classes
 - add Netty codec
 - add optional TLS handshake executor
+- add password modify operation
 - add server feature discovery
 - add thread-local executor
 - add "Who am I?" operation
 - receive and send TLS renegotiations
 - return TLS session
+- support all operational attributes
 
 Version 1.1.0:
 - add LDAP operations